services/auth/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.15.0
requests==2.25.0
sha256==0.1
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
